Move old JSONObject utility and test classes into org.json.*
