+++ api/src/org/json/old/JSONArrayDeserializer.java
@@ -1,4 +1,4 @@
-package org.labkey.api.util;
+package org.json.old;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
@@ -6,8 +6,6 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.fasterxml.jackson.databind.type.LogicalType;
 import com.fasterxml.jackson.databind.util.ClassUtil;
-import org.json.old.JSONArray;
-import org.json.old.JSONObject;
 
 import java.io.IOException;
