+++ api/src/org/json/old/JSONBaseSerializer.java
@@ -1,4 +1,4 @@
-package org.labkey.api.util;
+package org.json.old;
 
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
