+++ api/src/org/json/old/JSONObjectDeserializer.java
@@ -1,12 +1,10 @@
-package org.labkey.api.util;
+package org.json.old;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.fasterxml.jackson.databind.type.LogicalType;
-import org.json.old.JSONException;
-import org.json.old.JSONObject;
 
 import java.io.IOException;
