+++ api/src/org/labkey/api/util/JsonUtil.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.json.JsonOrgOldModule;
 
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -183,7 +184,7 @@ public class JsonUtil
         return result;
     }
 
-    // The new JSONObject.toMap() translates all JSONObjects into Maps and JSONArrays in Lists. In many cases, this is
+    // The new JSONObject.toMap() translates all JSONObjects into Maps and JSONArrays into Lists. In many cases, this is
     // fine, but some existing code paths want to maintain the contained JSONObjects and JSONArrays. This method does
     // that, acting more like the old JSONObject.toMap().
     public static void fillMapShallow(JSONObject json, Map<String, Object> map)
