+++ core/src/org/labkey/core/admin/createFolder.jsp
@@ -15,15 +15,15 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.json.old.JSONArray"%>
+<%@ page import="org.json.JSONArray"%>
 <%@ page import="org.labkey.api.data.ContainerManager" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
 <%@ page import="org.labkey.core.admin.AdminController" %>
-<%@ page import="org.labkey.core.portal.ProjectController" %>
 <%@ page import="org.labkey.core.portal.CollaborationFolderType" %>
+<%@ page import="org.labkey.core.portal.ProjectController" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%!
     @Override
