+++ core/src/org/labkey/core/admin/customizeMenu.jsp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 %>
-<%@ page import="org.json.old.JSONObject" %>
+<%@ page import="org.json.JSONObject" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
 <%@ page import="org.labkey.core.admin.CustomizeMenuForm" %>
@@ -36,7 +36,7 @@
 <script type="text/javascript" nonce="<%=getScriptNonce()%>">
     Ext.onReady(function() {
 
-        var bean = <%= text(new JSONObject(bean).toString()) %>;
+        var bean = <%=new JSONObject(bean)%>;
 
         Ext.QuickTips.init();
