+++ core/src/org/labkey/core/webdav/DavController.java
@@ -28,9 +28,9 @@ import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.json.old.JSONArray;
-import org.json.old.JSONObject;
-import org.json.old.JSONWriter;
+import org.json.JSONArray;
+import org.json.JSONObject;
+import org.json.JSONWriter;
 import org.junit.Assert;
 import org.junit.Test;
 import org.labkey.api.action.BaseViewAction;
