+++ query/src/org/labkey/query/olap/OlapTestCase.jsp
@@ -1,14 +1,12 @@
-<%@ page import="org.json.old.JSONObject" %>
+<%@ page import="org.json.JSONObject" %>
 <%@ page import="org.junit.Test" %>
 <%@ page import="org.labkey.api.security.User" %>
 <%@ page import="org.labkey.api.util.JunitUtil" %>
 <%@ page import="org.labkey.api.util.TestContext" %>
 <%@ page import="org.labkey.api.view.ActionURL" %>
-<%@ page import="static org.junit.Assert.*" %>
 <%@ page import="org.labkey.api.view.ViewServlet" %>
 <%@ page import="java.util.Map" %>
-<%@ page import="org.apache.logging.log4j.LogManager" %>
-<%@ page import="org.labkey.query.controllers.OlapController" %>
+<%@ page import="static org.junit.Assert.*" %>
 <%@ page extends="org.labkey.api.jsp.JspTest.DRT" %>
 
 <%!
@@ -31,11 +29,9 @@ boolean canExecuteMdx(String configId) throws Exception
     config.put("configId", configId);
     config.put("schemaName", "OlapTest");
     JSONObject result = executeJsonApi(new ActionURL("olap","executeMdx", JunitUtil.getTestContainer()), TestContext.get().getUser(), config.toString());
-    if (null != result.get("success") && Boolean.FALSE == result.getBoolean("success"))
+    if (!result.optBoolean("success", true))
         return false;
-    if (null == result.get("cells"))
-        return false;
-    return true;
+    return null != result.get("cells");
 }
 
 boolean canExecuteJson(String configId) throws Exception
@@ -47,11 +43,9 @@ boolean canExecuteJson(String configId) throws Exception
     config.put("cubeName", "Facts");
 
     JSONObject result = executeJsonApi(new ActionURL("olap","jsonQuery", JunitUtil.getTestContainer()), TestContext.get().getUser(), config.toString());
-    if (null != result.get("success") && Boolean.FALSE == result.getBoolean("success"))
+    if (!result.optBoolean("success", true))
         return false;
-    if (null == result.get("cells"))
-        return false;
-    return true;
+    return null != result.get("cells");
 }
 
 boolean canExecuteCountDistinct(String configId) throws Exception
@@ -63,11 +57,9 @@ boolean canExecuteCountDistinct(String configId) throws Exception
     config.put("cubeName", "Facts");
 
     JSONObject result = executeJsonApi(new ActionURL("olap","countDistinctQuery", JunitUtil.getTestContainer()), TestContext.get().getUser(), config.toString());
-    if (null != result.get("success") && Boolean.FALSE == result.getBoolean("success"))
+    if (!result.optBoolean("success", true))
         return false;
-    if (null == result.get("cells"))
-        return false;
-    return true;
+    return null != result.get("cells");
 }
 
 @Test
